Added unified cdc update tests.

+++ single-module/_build-and-test-all-eventuate-local-postgres.sh
@@ -1,5 +1,12 @@
 #! /bin/bash
 
+NO_RM=false
+
+if [ "$1" = "--no-rm" ] ; then
+  NO_RM=true
+  shift
+fi
+
 . ./set-env-postgres.sh
 
 docker-compose -f docker-compose-eventuate-local-postgres-${MODE}.yml down
@@ -13,4 +20,6 @@ docker-compose -f docker-compose-eventuate-local-postgres-${MODE}.yml up --build
 
 ./gradlew cleanTest e2eTest -P ignoreE2EFailures=false
 
-docker-compose -f docker-compose-eventuate-local-postgres-${MODE}.yml down
+if [ $NO_RM = false ] ; then
+  docker-compose -f docker-compose-eventuate-local-postgres-${MODE}.yml down
+fi
